\name{ReturnSeries_data}
\alias{ReturnSeries_data}
\docType{data}
\title{Return Series Data}
\description{
Monthly return series of international soybean price start from January 1980.
}
\usage{data("ReturnSeries_data")}
\format{
  A data frame with 107 observations on the following variable.

}
\details{
Dataset contain 107 0bservations of monthly return series of International soyabean price. It is obtained from World Bank "Pink sheet"
}
\source{
https://www.worldbank.org/en/research/commodity-markets
}
\references{
https://www.worldbank.org/en/research/commodity-markets
}
\examples{
data(ReturnSeries_data)
## maybe str(ReturnSeries_data) ; plot(ReturnSeries_data) ...
}
\keyword{datasets}
